refactor: rename outward data processing module and update imports
